Copy in scripts from emews-r-x

+++ conda-build.sh
@@ -12,36 +12,80 @@ set -eu
 #     specified in meta.yaml:source:path:
 
 # Hard-coded for now- on other systems, we use community R packages
-PLATFORM=osx-arm64
+# Also set in meta.yaml
+export PLATFORM=${PLATFORM:-osx-arm64}
 
 help()
 {
   cat <<END
 
 Options:
-   -C configure-only
-
+   -S Settings:  Just reports startup settings and exits
+   -C Configure: Just runs configure, not make
 END
 }
 
-C=""
-zparseopts -D -E -F h=HELP C=C
+C="" S=""
+zparseopts -D -E -F h=HELP C=C S=S
 
 if (( ${#HELP} )) {
   help
   exit
 }
 
+if (( ${#C} )) export CONFIG_ONLY=1
+
 # Get this directory
 THIS=${0:A:h}
+
+if (( ${#*} != 1 )) {
+  print "Provide R SVN location!"
+  return 1
+}
+# This is substituted into meta.yaml:
+export R_SVN=$1
+
+if [[ ! -d $R_SVN ]] {
+  print "does not exist: R_SVN=$R_SVN"
+  return 1
+}
+
+if ! grep -q "R_HOME" $R_SVN/Makefile.in
+then
+  print "wrong directory: R_SVN=$R_SVN"
+  return 1
+fi
+
+if [[ $PLATFORM =~ osx-* ]] {
+  if [[ $(uname) != "Darwin" ]] {
+    print "platform mismatch!"
+    print "uname: $(uname)"
+    print "PLATFORM=$PLATFORM"
+    return 1
+  }
+}
 
 DATE_FMT_S="%D{%Y-%m-%d} %D{%H:%M:%S}"
 log()
 # General-purpose log line
-# You may set global LOG_LABEL to get a message prefix
 {
-  print ${(%)DATE_FMT_S} ${LOG_LABEL:-} ${*}
+  print ${(%)DATE_FMT_S} "conda-build.sh:" ${*}
 }
+
+if [[ ${GITHUB_ACTIONS:-0} == true ]]
+then
+  source ./enable-python.sh
+fi
+
+# Look up executable:
+PYTHON_EXE=( =python )
+# Get its directory:
+PYTHON_BIN=${PYTHON_EXE:h}
+
+# print LISTING
+# print PYTHON_EXE $PYTHON_EXE $PYTHON_BIN
+# ls $PYTHON_BIN
+# conda list
 
 # Check that the conda-build tool in use is in the
 #       selected Python installation
@@ -55,20 +99,11 @@ fi
 CONDA_BUILD_TOOL=( =conda-build )
 # Get its directory:
 TOOLDIR=${CONDA_BUILD_TOOL:h}
-# Look up executable:
-PYTHON_EXE=( =python )
-# Get its directory:
-PYTHON_BIN=${PYTHON_EXE:h}
 if [[ ${TOOLDIR} != ${PYTHON_BIN} ]] {
   log "conda-build is not in your python directory!"
   log "            this is probably wrong!"
   log "            run ./setup-conda.sh"
   return 1
-}
-
-if (( ${#C} )) {
-  log "configure-only: exit."
-  exit
 }
 
 # Backup the old log
@@ -79,31 +114,49 @@ if [[ -f $LOG ]] {
   print
 }
 
+if (( ${#S} )) {
+  log "settings-only: exit." | tee $LOG
+  return
+}
+
 {
-  log "CONDA BUILD: START: ${(%)DATE_FMT_S}"
+  log "CONDA BUILD: START:"
   print
   (
     log "using python: " $( which python )
     log "using conda:  " $( which conda  )
+    log "PLATFORM:     " $PLATFORM
     print
     conda env list
     print
 
-    set -x
     # This purge-all is extremely important:
+    log "purge-all ..."
     conda build purge-all
 
+    BUILD_ARGS=(
+      -c conda-forge
+      --dirty
+      --keep-old-work
+      # --no-remove-work-dir
+      --prefix-length 128
+      .
+    )
+
     # Build the package!
-    conda build \
-          -c conda-forge \
-          --dirty \
-          .
+    log "main build args: $BUILD_ARGS"
+    conda build $BUILD_ARGS
   )
   log "CONDA BUILD: STOP: ${(%)DATE_FMT_S}"
 } |& tee $LOG
 print
-log "conda build succeeded."
+log "SUCCESS."
 print
+
+if (( ${CONFIG_ONLY:-0} )) {
+  log "configure-only: done."
+  return
+}
 
 # Look for success from meta.yaml:test:commands:
 # Output
@@ -141,3 +194,6 @@ checksum()
   log $T
   log "HASH:" $( checksum $PKG )
 ) | tee -a $LOG
+
+# This is mostly for GitHub Action step checking
+echo "CONDA-BUILD: SUCCESS" > $LOG
